google and github oauth done : AUTH COMPLETE
